docker_compose/app/requirements.txt
uwsgi==2.0.20
django==3.2.*
flake8==4.0.1
python-dotenv==0.21.0
django-split-settings==1.1.0
psycopg2==2.9.3
Faker==13.15.1
django-debug-toolbar==3.5.0
django-extensions
django-cors-headers
django-probes